fastapi-react/backend/requirements.txt
uvicorn
fastapi
python-multipart
pydantic
passlib
bcrypt
python-jose
sqlalchemy
psycopg2-binary
python-dotenv
alembic
redis
aioredis